Handle OOM Errors better
On OOM error, we can't load a module because we are out of memory.  We
used to just crash.  Now we preload the module so in most cases, we
can get coherent error messages even when we are out of memory.  We
also handle failures better when we are unable to allocate something
during error handling.